Fix update entities fields

+++ ui/auto/libperl-OBM/OBM/DbUpdater/groupUpdater.pm
@@ -235,6 +235,15 @@ sub _delete {
             return 1;
         }
 
+        $query = 'DELETE FROM P_field
+                    WHERE entity_id=(SELECT groupentity_entity_id
+                                                    FROM GroupEntity
+                                                    WHERE groupentity_group_id = '.$entity->getId().')';
+        if( !defined( $dbHandler->execQuery( $query, \$sth ) ) ) {
+            $self->_log( 'problème à la mise à jour BD '.$entity->getDescription(), 1 );
+            return 1;
+        }
+
         $query = 'DELETE FROM P_GroupEntity WHERE groupentity_group_id='.$entity->getId();
         if( !defined( $dbHandler->execQuery( $query, \$sth ) ) ) {
             $self->_log( 'problème à la mise à jour BD '.$entity->getDescription(), 1 );
