+++ ui/auto/libperl-OBM/OBM/DbUpdater/hostUpdater.pm
@@ -231,6 +231,15 @@ sub _delete {
             return 1;
         }
 
+        $query = 'DELETE FROM P_field
+                    WHERE entity_id=(SELECT hostentity_entity_id
+                                                    FROM HostEntity
+                                                    WHERE hostentity_host_id = '.$entity->getId().')';
+        if( !defined( $dbHandler->execQuery( $query, \$sth ) ) ) {
+            $self->_log( 'problème à la mise à jour BD '.$entity->getDescription(), 1 );
+            return 1;
+        }
+
         $query = 'DELETE FROM P_ServiceProperty
                     WHERE serviceproperty_entity_id=(SELECT hostentity_entity_id
                                                     FROM HostEntity
